Add handlers for rest of the commands
