+++ eigentrust-cli/src/main.rs
@@ -31,10 +31,10 @@ mod fs;
 use clap::Parser;
 use cli::*;
 use dotenv::dotenv;
-use eigentrust::{error::EigenError, eth::deploy_as, Client, ClientConfig};
+use eigentrust::{error::EigenError, ClientConfig};
 use env_logger::{init_from_env, Env};
-use fs::{load_config, load_mnemonic};
-use log::{debug, info};
+use fs::load_config;
+use log::info;
 
 #[tokio::main]
 async fn main() -> Result<(), EigenError> {
@@ -43,22 +43,10 @@ async fn main() -> Result<(), EigenError> {
 	let mut config: ClientConfig = load_config()?;
 
 	match Cli::parse().mode {
-		Mode::Attest(attest_data) => {
-			let attestation = attest_data.to_attestation_raw(&config)?;
-			debug!("Attesting:{:?}", attestation);
-
-			let mnemonic = load_mnemonic();
-			let client = Client::new(config, mnemonic);
-			client.attest(attestation).await?;
-		},
+		Mode::Attest(attest_data) => handle_attest(config, attest_data).await?,
 		Mode::Attestations => handle_attestations(config).await?,
 		Mode::Bandada(bandada_data) => handle_bandada(&config, bandada_data).await?,
-		Mode::Deploy => {
-			let mnemonic = load_mnemonic();
-			let client = Client::new(config, mnemonic);
-			let as_address = deploy_as(client.get_signer()).await?;
-			info!("AttestationStation deployed at {:?}", as_address);
-		},
+		Mode::Deploy => handle_deploy(config).await?,
 		Mode::LocalScores => handle_scores(config, AttestationsOrigin::Local).await?,
 		Mode::Proof => handle_proof(config).await?,
 		Mode::Scores => handle_scores(config, AttestationsOrigin::Fetch).await?,
